+++ hume-scripting/src/builtins/mod.rs
@@ -413,7 +413,8 @@ pub(crate) fn register_all(steel: &mut Engine) {
         cmd  "lsp-server-for-buffer" lsp::lsp_server_for_buffer(bid: args::BidArg);
         open "lsp-registered-for-language?" lsp::lsp_registered_for_language(language: SteelVal);
         cmd "lsp-position-params" lsp::lsp_position_params(bid: args::BidArg);
-        cmd "lsp-range-params" lsp::lsp_range_params(bid: args::BidArg);
+        cmd "lsp-primary-range-params" lsp::lsp_primary_range_params(bid: args::BidArg);
+        cmd "lsp-selections-range-params" lsp::lsp_selections_range_params(bid: args::BidArg);
         cmd "lsp-position->offset" lsp::lsp_position_to_offset(bid: args::BidArg, position: SteelVal);
         cmd "lsp-range->offsets" lsp::lsp_range_to_offsets(bid: args::BidArg, range: SteelVal);
         cmd "lsp-label-offsets->text" lsp::lsp_label_offsets_to_text(bid: args::BidArg, label: SteelVal, offsets: SteelVal);